requirements.txt
google-generativeai
streamlit
psycopg2-binary
python-dotenv
requests